Windows 'int' is 32 bits, correct size_t and ms types
- All stdcxx size and count results are expressed as size_t
- Microseconds as 45 bits will always be a 64 bit int

+++ source/server/guarddog_impl.h
@@ -72,7 +72,7 @@ public:
   /**
    * Exposed for testing purposes only (but harmless to call):
    */
-  int loopIntervalForTest() const { return loop_interval_.count(); }
+  int64_t loopIntervalForTest() const { return loop_interval_.count(); }
 
   /**
    * Test hook to force a step() to catch up with the current simulated
