+++ source/server/http/admin.cc
@@ -1457,7 +1457,7 @@ bool AdminImpl::addHandler(const std::string& prefix, const std::string& help_te
 }
 
 bool AdminImpl::removeHandler(const std::string& prefix) {
-  const uint size_before_removal = handlers_.size();
+  const size_t size_before_removal = handlers_.size();
   handlers_.remove_if(
       [&prefix](const UrlHandler& entry) { return prefix == entry.prefix_ && entry.removable_; });
   if (handlers_.size() != size_before_removal) {
